Separated the controllers into C/Q services and routes into commands and queries
