+++ src/routes/products-queries-routes.js
@@ -3,10 +3,10 @@ const express = require('express');
 const { check } = require('express-validator');
 
 /* Application files */
-const productsQueries = require('../services/queries/products-queries');
+const productsQueriesController = require('../services/queries/products-queries');
 
 const router = express.Router();
 
-router.get('/', productsQueries.getAllProducts);
+router.get('/', productsQueriesController.getAllProducts);
 
 module.exports = router;
