+++ src/routes/stock-commands-routes.js
@@ -3,7 +3,7 @@ const express = require('express');
 const { check } = require('express-validator');
 
 /* Application files */
-const stockCommands = require('../services/commands/stock-commands');
+const stockCommandsController = require('../services/commands/stock-commands');
 
 const router = express.Router();
 
@@ -11,12 +11,12 @@ router.post('/:id/restock',
     [
         check('stock').not().isEmpty()
     ],
-    stockCommands.restock);
+    stockCommandsController.restock);
 
 router.post('/:id/sell', 
     [
         check('stock').not().isEmpty()
     ],
-    stockCommands.sell);
+    stockCommandsController.sell);
 
 module.exports = router;
